Move processRevisions to util
