+++ indico/modules/events/editing/client/js/editing/timeline/__tests__/selectors.test.js
@@ -7,8 +7,7 @@
 
 import {FinalRevisionState, InitialRevisionState} from 'indico/modules/events/editing/models';
 
-import {processRevisions} from '../selectors';
-import {revisionStates} from '../util';
+import {processRevisions, revisionStates} from '../util';
 
 describe('timeline selectors', () => {
   it('should describe each revision transition', () => {
